+++ sh4rewithme-perf/src/test/scala/users/NewUser.scala
@@ -13,8 +13,8 @@ class NewUser extends Simulation {
 	val extHost = Option(System.getProperty("extHost")).getOrElse("localhost")
 	val extPort = Integer.getInteger("extPort", 9090)
 	val extUsers = Integer.getInteger("extUsers", 1)
-	val extRampup = Integer.getInteger("extRampup", 0).toLong
-	val extPause = Integer.getInteger("extPause", 1).toLong
+	val extRampup = java.lang.Long.getLong("extRampup", 0L)
+	val extPause = java.lang.Long.getLong("extPause", 1L)
 	val extLoop = Integer.getInteger("extLoop", 100)
 
 	val extBaseUrl = if (extPort == 443)
